CREP-29 Move post related components
Moved post related components to the new components folder, and fixed imports relating to those components
